+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 ## 0.11
 ### Added
 - Filter and sort functionality on _Entity Type List_
+- Dependencies added to _System Separator_ attribute
 ### Fixed
 - Missing "Color Dots" of _Entity Type_ (colors are now handled on the Entity Type itself and do no longer rely on the _Map_ Plugin)
 - 'Jumping' behavior on floating quick access controls on the _Entity Type List_ (copy, duplicate, delete)
 - Attribute Dependency inverted
+### Changed
+- Users can add more than one dependency for an attribute and use groups and union/intersect operators
 
 ## 0.10.2
 ### Fixed